requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
supabase==2.22.0
python-dotenv==1.0.0
httpx==0.27.0
fastapi
uvicorn
httpx
